Made autoinstall, clone, destroy VMs

+++ destroy_clones.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
+domen=$1
+name=${domen}"-clone"
 for x in $(virsh list --all --name);
 do
+if grep -q "$name" <<< $x ; then
 virsh destroy $x;
 virsh undefine $x --remove-all-storage;
+fi
 done
